fix: polish Method page typography, spacing, and scroll animations
- Add Instrument Serif font for principle titles (more editorial feel)
- Widen layout from max-w-3xl to max-w-4xl for breathing room
- Increase title size to 8xl on desktop
- Make PHILOSOPHY label more subtle (smaller, reduced opacity)
- Improve body text line-height (1.85) and letter-spacing
- Add warm muted text color instead of cold gray
- Add border-left accent treatment for principle labels
- Enlarge principle titles (5xl) and subtitles (xl)
- Replace thin hr dividers with gradient-fade dividers
- Increase spacing rhythm throughout (24/32 units)
- Add framer-motion scroll-triggered fade-in for all sections
- Increase closing section visual weight (2xl text, more spacing)
